Year switching was fixed
The bug in week view when the month is January and the previous week go 
to previus year is fixed.
The same bug was fixed when the month is December and the next week go 
to the next year. 

Reason:
The if statement in the week.component constructor checks the parameters 
passed via url and, if are not all present, initialize the week on the 
current date.
But the native Date class counts January as month number 0 month that 
cause the logical expression result contained in the if statement to be 
always false.
Resolution:
Not so elegant but for now the if statement checks only the parameters 
relative to day and year.